requirements.txt
asgiref==3.8.1
astroid==3.3.9
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coverage==7.8.0
cryptography==46.0.5
cssbeautifier==1.15.4
cssselect2==0.8.0
dill==0.3.9
Django==5.2
django-dump-load-utf8==0.0.4
django-phonenumber-field==8.1.0
django-simple-history==3.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djlint==1.36.4
drf-spectacular==0.28.0
EditorConfig==0.17.0
et_xmlfile==2.0.0
fonttools==4.58.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
isort==6.0.1
jsbeautifier==1.15.4
json5==0.12.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.4
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
phonenumbers==9.0.3
pillow==11.2.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.3.6
pypdf==5.6.0
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
robotframework==7.3
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-seleniumlibrary==6.7.1
rpds-py==0.26.0
ruff==0.11.4
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zopfli==0.2.3.post1
